DOMS: scripts, lecture notes, our paper and updated readme
